requirements.txt
APScheduler==3.3.1
beautifulsoup4==4.5.1
click==6.6
contextlib2==0.5.4
decorator==4.0.11
Flask==0.12
funcsigs==1.0.2
future==0.16.0
futures==3.0.5
geocoder==1.19.0
google-api-python-client==1.5.5
Google-Search-API==1.1.13
hanziconv==0.3.2
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8.1
line-bot-sdk==1.0.2
MarkupSafe==0.23
mock==2.0.0
numpy==1.11.3
oauth2client==4.0.0
pandas==0.19.2
pbr==1.10.0
pprint==0.1
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pymongo==3.4.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
ratelim==0.1.6
requests==2.12.4
rsa==3.4.2
selenium==2.53.6
six==1.10.0
tqdm==4.10.0
tzlocal==1.3
ua-parser==0.7.2
Unidecode==0.4.19
uritemplate==3.0.0
vcrpy==1.10.4
Werkzeug==0.11.15
wrapt==1.10.8
